requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0.0
accelerate>=1.10.0
timm>=1.0.19
triton>=3.4.0
sentencepiece>=0.2.1
